requirements.txt
paho-mqtt==1.6.1
PyYAML==6.0.1
pygame>=2.5.2
numpy==1.26.4
pyaudio>=0.2.14
